Revert "Added a Level class and changed FPS to be stored as an Integer."

+++ Source/Game.h
@@ -34,7 +34,7 @@ private:
 
 	std::queue<double> FrameQueue;
 
-	int FrameRate;
+	float FrameRate;
 
 public:
 	//CTORS
